Split namespace Json into Time and Numbers
