+++ src/Json/Numbers/SingleConv.cs
@@ -2,10 +2,10 @@ using System.Globalization;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace SurrealDB.Json;
+namespace SurrealDB.Json.Numbers;
 
-public sealed class DoubleConv : JsonConverter<double> {
-    public override double Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) {
+public sealed class SingleConv : JsonConverter<float> {
+    public override float Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) {
         if (reader.TokenType is JsonTokenType.Null or JsonTokenType.None) {
             return default;
         }
@@ -16,26 +16,26 @@ public sealed class DoubleConv : JsonConverter<double> {
                 return default;
             }
 
-            double v = SpecialNumbers.ToDouble(str);
+            float v = SpecialNumbers.ToSingle(str);
             if (v != default) {
                 return v;
             }
 
-            return Double.Parse(str, NumberStyles.Float, NumberFormatInfo.InvariantInfo);
+            return Single.Parse(str, NumberStyles.Float, NumberFormatInfo.InvariantInfo);
         }
 
         if (reader.TokenType == JsonTokenType.Number) {
-            return reader.GetDouble();
+            return reader.GetSingle();
         }
 
         throw new JsonException("Could not parse the number.");
     }
 
-    public override double ReadAsPropertyName(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) {
+    public override float ReadAsPropertyName(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) {
         return Read(ref reader, typeToConvert, options);
     }
 
-    public override void Write(Utf8JsonWriter writer, double value, JsonSerializerOptions options) {
+    public override void Write(Utf8JsonWriter writer, float value, JsonSerializerOptions options) {
         string? str = SpecialNumbers.ToSpecial(in value);
         if (str is not null) {
             writer.WriteStringValue(str.AsSpan());
@@ -45,7 +45,7 @@ public sealed class DoubleConv : JsonConverter<double> {
         writer.WriteNumberValue(value);
     }
 
-    public override void WriteAsPropertyName(Utf8JsonWriter writer, double value, JsonSerializerOptions options) {
+    public override void WriteAsPropertyName(Utf8JsonWriter writer, float value, JsonSerializerOptions options) {
         string? str = SpecialNumbers.ToSpecial(in value);
         if (str is not null) {
             writer.WritePropertyName(str.AsSpan());
